Move MCP server implementation into the mcp domain
Constraint: src/mcp-server imports are broad across tools, tests, and runtime entrypoints, so the root path remains a compatibility shim.\nRejected: Rewriting every MCPServer caller in the same PR | it would produce broad churn and distract from the ownership move.\nRejected: Keeping the implementation at the source root | it leaves protocol/server internals outside the existing src/mcp domain.\nConfidence: high\nScope-risk: moderate\nDirective: Keep new MCP protocol/server internals under src/mcp; migrate callers away from the shim in smaller batches.\nTested: npx jest --config jest.ci.config.js --runInBand --silent --runTestsByPath tests/tools/manifest-only-registration.test.ts tests/mcp/server.test.ts tests/mcp/server-profile.test.ts tests/mcp/server-shutdown.test.ts tests/mcp/server-abort-accounting.test.ts tests/mcp/request-ingress.test.ts tests/metrics/tool-output-observability.test.ts tests/tools/journal.test.ts tests/cli/entrypoint-parity.test.ts\nTested: npm run build\nTested: npm run lint\nTested: npm run lint:tier\nTested: git diff --check\nNot-tested: Full Jest suite.
